This commit contains the simplest possible demonstration of the bug which occurs when: a factory function injected by Factory.of() instantiates a class and a dynamic dependency is provided to the factory function and the instantiated class has the @autoinject() decorator or the instantiated class has ANY decorator according to docs:
Source:
http://aurelia.io/docs.html#/aurelia/dependency-injection/1.0.0-beta.1.2.3/doc/article/dependency-injection-basics

"Interestingly, you don't need to use our autoinject decorator at all to get the above to work. The TypeScript compiler will emit the type metadata if any decorator is added to the class. Aurelia can read this metadata regardless of what decorator triggers TypeScript to add it. We simply provide the autoinject decorator for consistency and clarity."

Cause:
This is because function invokeWithDynamicDependencies(container, fn, staticDependencies, dynamicDependencies) is used to provide the dependencies to the instantiated class, and applies all static dependencies to the result array in addition to all dynamic dependencies, when it should stop populating the array with static dependencies once the size of the array reaches the number of constructor parameters minus the number of dynamic dependencies. Currently the number of constructor parameters is not exposed within the function.